django/requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-cors-headers
psycopg2
django-polymorphic
Pillow
pylint
pylint-django
